requirements.txt
Django==5.1.6
django-widget-tweaks==1.4.8
Pillow==8.3.2
psycopg2-binary==2.9.1
gunicorn==20.1.0
whitenoise==5.3.0
django-crispy-forms==1.12.0
django-allauth==0.45.0
django-ckeditor==6.1.0
django-autoslug-1.9.9